signed/unsigned comparison warning fix
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@29536 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/intl.cpp
@@ -1126,7 +1126,7 @@ bool wxMsgCatalogFile::Load(const wxChar *szDirPrefix, const wxChar *szName0,
 
   // read the whole file in memory
   m_pData = new size_t8[nSize];
-  if ( fileMsg.Read(m_pData, nSize) != nSize ) {
+  if ( fileMsg.Read(m_pData, nSize) != (size_t)nSize ) {
     wxDELETEA(m_pData);
     return false;
   }
